requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
idna==3.4
kombu==5.2.4
Pillow==9.2.0
prompt-toolkit==3.0.31
PyJWT==2.5.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
